+++ R/bootstrap.R
@@ -19,6 +19,28 @@ bootstrap <- function(version, library) {
 
 }
 
+renv_bootstrap_download_impl <- function(url, destfile) {
+
+  mode <- "wb"
+
+  # https://bugs.r-project.org/bugzilla/show_bug.cgi?id=17715
+  fixup <-
+    Sys.info()[["sysname"]] == "Windows" &&
+    identical(getOption("download.file.method"), "wininet") &&
+    substring(url, 1, 5) == "file:"
+
+  if (fixup)
+    mode <- "w+b"
+
+  download.file(
+    url      = url,
+    destfile = destfile,
+    mode     = mode,
+    quiet    = TRUE
+  )
+
+}
+
 renv_bootstrap_download <- function(version) {
 
   methods <- list(
@@ -77,7 +99,7 @@ renv_bootstrap_download_cran_archive <- function(version) {
   for (url in urls) {
 
     status <- tryCatch(
-      download.file(url, destfile, mode = "wb", quiet = TRUE),
+      renv_bootstrap_download_impl(url, destfile),
       condition = identity
     )
 
@@ -122,7 +144,7 @@ renv_bootstrap_download_github <- function(version) {
   destfile <- file.path(tempdir(), name)
 
   status <- tryCatch(
-    download.file(url, destfile = destfile, mode = "wb", quiet = TRUE),
+    renv_bootstrap_download_impl(url, destfile),
     condition = identity
   )